pyproject.toml
[project]
name = "capacity-ar-api"
version = "0.1.0"
description = "MVP backend API for the Capacity AR learning platform."
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "sqlalchemy==2.0.36",
    "psycopg[binary]==3.2.13",
    "pydantic-settings==2.7.1",
    "email-validator==2.2.0",
    "google-generativeai==0.8.6",
    "groq==0.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "httpx==0.28.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
